Update Betav2b_MacOSX_patch.md
Hi this is the first time I try using github. I just meant to change the numer 4 into the number 6 (amount of files in hte patch).
But then I found myself changing the order of the information in the instructions, because I have really been puzzling about things I encounterd while trying to install the patch (for example the apple warning about zsh). I thought it might be usefull to mention this before one ecounters it. The same goes for the other things I have dragged more to the beginning.